Changed Price object to store its number values as double primitives in order to obtain correct values during conversion calculations when using values retrieved from the database

+++ src/main/java/com/alexsnowm/grocerycomparer/models/forms/CreateItemForm.java
@@ -12,7 +12,7 @@ public class CreateItemForm {
     private String itemName;
 
     @Digits(integer = 7, fraction = 2, message = "Enter number to no more than 2 decimal places")
-    private BigDecimal priceNumber;
+    private Double priceNumber;
 
     private ItemMeasure measure;
     private String priceAisle;
@@ -29,11 +29,11 @@ public class CreateItemForm {
         this.itemName = itemName;
     }
 
-    public BigDecimal getPriceNumber() {
+    public Double getPriceNumber() {
         return priceNumber;
     }
 
-    public void setPriceNumber(BigDecimal priceNumber) {
+    public void setPriceNumber(Double priceNumber) {
         this.priceNumber = priceNumber;
     }
